+++ example/lib/ui/zego_ui_tool.dart
@@ -20,4 +20,25 @@ class ZegoUITool {
     });
   }
 
+  static showImage(BuildContext context, MemoryImage imageData) {
+    showDialog(context: context, builder: (BuildContext context) {
+      if (imageData == null) {
+        print('[ZegoUITool][showImage] image data is null!');
+        return null;
+      }
+
+      return AlertDialog(
+        content: Image(image: imageData, width: MediaQuery.of(context).size.width / 2, height: MediaQuery.of(context).size.width / 2),
+        actions: <Widget>[
+          FlatButton(
+            child: Text('OK'),
+            onPressed: () {
+              Navigator.of(context).pop();
+            },
+          )
+        ],
+      );
+    });
+  }
+
 }